+++ src/jk_cmdoutputparsinghelper/ColumnDef.py
@@ -15,14 +15,16 @@ class ColumnDef(object):
 	#
 	# @param		str name				(required) The name of the column.
 	# @param		callable valueParser	(optional) A value parser.
+	# @param		str typeName			(optional) The name of the type.
 	#
-	def __init__(self, name:str, valueParser = None):
+	def __init__(self, name:str, valueParser = None, typeName = None):
 		assert isinstance(name, str)
 		if valueParser is not None:
 			assert callable(valueParser)
 
 		self.name = name
 		self.valueParser = valueParser
+		self.typeName = typeName
 	#
 
 	################################################################################################################################
@@ -36,6 +38,30 @@ class ColumnDef(object):
 	################################################################################################################################
 	## Public Methods
 	################################################################################################################################
+
+	def __str__(self):
+		_typeName = None
+		if self.typeName is not None:
+			_typeName = self.typeName
+		else:
+			if isinstance(self.valueParser, type):
+				_typeName = self.valueParser.__name__
+			else:
+				_typeName = "???"
+		return self.name + ":" + _typeName
+	#
+
+	def __repr__(self):
+		_typeName = None
+		if self.typeName is not None:
+			_typeName = self.typeName
+		else:
+			if isinstance(self.valueParser, type):
+				_typeName = self.valueParser.__name__
+			else:
+				_typeName = "???"
+		return self.name + ":" + _typeName
+	#
 
 #