+++ src/jk_cmdoutputparsinghelper/__init__.py
@@ -1,7 +1,7 @@
 ﻿
 
 
-__version__ = "0.2021.2.17.1"
+__version__ = "0.2021.6.6"
 
 
 
@@ -15,4 +15,3 @@ from .TextData import TextData
 from .ValueParserDef import ValueParserDef
 from .LineParser_ParseAtFirstDelimiter import LineParser_ParseAtFirstDelimiter
 from .ValueParser_ByteWithUnit import ValueParser_ByteWithUnit
-