+++ HISTORY.md
@@ -6,3 +6,6 @@
 	* Added: HISTORY.md
 	* Improved: TextData to handle empty content more precisely
 	* Added: unit test for TextData
+
+* 2021-06-06
+	* Improved: added more convenience methods to LineList and Table